fix: address all the nitpicks from @taghizadlaura

+++ app/src/main/java/com/android/periodpals/ui/components/ProfileComponents.kt
@@ -235,7 +235,7 @@ fun ProfileSaveButton(
                 dob = dobState.value,
                 description = descriptionState.value,
                 imageUrl = profileImageState.value,
-                preferredDistance = preferredDistance.toInt(),
+                preferredDistance = preferredDistance,
             )
         userViewModel.saveUser(
             user = newUser,
